feat: created gig schema and collaboration schema to handle the collabs between influencers
